add first draft

+++ TagsCloudVisualization/Config.cs
@@ -5,8 +5,8 @@ namespace TagsCloudVisualization
     public struct Config
     {
         public Size Size { get; set; }
-        public int WordCountToStatistic { get; set; }
-        public double Density { get; set; }
+        public uint WordCountToStatistic { get; set; }
+        public uint Density { get; set; }
         public byte MinWordToStatisticLength { get; set; }
         public string Font { get; set; }
         public Color? Color { get; set; }
